Move new serialiers/viewsets to v2
These are only necessary for v2 API.
