+++ d4s2_api/serializers.py
@@ -85,28 +85,3 @@ class UserSerializer(serializers.HyperlinkedModelSerializer):
         fields = ('id', 'user_id', 'url', 'dds_id', 'full_name', 'email')
 
 
-class DDSUserSerializer(serializers.Serializer):
-    """
-    Serializer for DukeDS users API
-    """
-    id = serializers.UUIDField()
-    username = serializers.CharField()
-    full_name = serializers.CharField()
-    first_name = serializers.CharField()
-    last_name = serializers.CharField()
-    email = serializers.CharField()
-
-    class Meta:
-        resource_name = 'duke-ds-users'
-
-
-class DDSProjectSerializer(serializers.Serializer):
-    """
-    Serializer for DukeDS projects API
-    """
-    id = serializers.UUIDField()
-    name = serializers.CharField()
-    description = serializers.CharField()
-
-    class Meta:
-        resource_name = 'duke-ds-projects'
