+++ d4s2_api/views.py
@@ -2,26 +2,11 @@ from rest_framework import viewsets, status, permissions
 from rest_framework.exceptions import APIException, ValidationError
 from rest_framework.decorators import detail_route
 from d4s2_api.models import Delivery, Share
-from d4s2_api.serializers import DeliverySerializer, ShareSerializer, DDSUserSerializer, DDSProjectSerializer
-from switchboard.dds_util import DDSUser, DDSProject
+from d4s2_api.serializers import DeliverySerializer, ShareSerializer
 from d4s2_api.utils import ShareMessage, DeliveryMessage
 from switchboard.dds_util import DDSUtil
 from django.core.urlresolvers import reverse
 from django_filters.rest_framework import DjangoFilterBackend
-
-
-class DataServiceUnavailable(APIException):
-    status_code = 503
-    default_detail = 'Data Service temporarily unavailable, try again later.'
-
-
-class WrappedDataServiceException(APIException):
-    """
-    Converts error returned from DukeDS python code into one appropriate for django.
-    """
-    def __init__(self, data_service_exception):
-        self.status_code = data_service_exception.status_code
-        self.detail = data_service_exception.message
 
 
 class AlreadyNotifiedException(APIException):
@@ -86,51 +71,3 @@ class ShareViewSet(viewsets.ModelViewSet):
         message.send()
         share.mark_notified(message.email_text)
         return self.retrieve(request)
-
-
-class DDSViewSet(viewsets.ReadOnlyModelViewSet):
-    permission_classes = (permissions.IsAuthenticated,)
-
-    def _ds_operation(self, func, *args):
-        try:
-            return func(*args)
-        except WrappedDataServiceException:
-            raise # passes along status code, e.g. 404
-        except Exception as e:
-            raise DataServiceUnavailable(e)
-
-
-class DDSUsersViewSet(DDSViewSet):
-    """
-    Interfaces with DukeDS API to provide project listing and details.
-    Though it is not backed by django models, the ReadOnlyModelViewSet base class
-    still works well
-    """
-    serializer_class = DDSUserSerializer
-
-    def get_queryset(self):
-        dds_util = DDSUtil(self.request.user)
-        return self._ds_operation(DDSUser.fetch_list, dds_util)
-
-    def get_object(self):
-        dds_user_id = self.kwargs.get('pk')
-        dds_util = DDSUtil(self.request.user)
-        return self._ds_operation(DDSUser.fetch_one, dds_util, dds_user_id)
-
-
-class DDSProjectsViewSet(DDSViewSet):
-    """
-    Interfaces with DukeDS API to provide project listing and details.
-    Though it is not backed by django models, the ReadOnlyModelViewSet base class
-    still works well
-    """
-    serializer_class = DDSProjectSerializer
-
-    def get_queryset(self):
-        dds_util = DDSUtil(self.request.user)
-        return self._ds_operation(DDSProject.fetch_list, dds_util)
-
-    def get_object(self):
-        dds_user_id = self.kwargs.get('pk')
-        dds_util = DDSUtil(self.request.user)
-        return self._ds_operation(DDSProject.fetch_one, dds_util, dds_user_id)
