+++ d4s2_api_v2/serializers.py
@@ -1,0 +1,28 @@
+from rest_framework import serializers
+
+
+class DDSUserSerializer(serializers.Serializer):
+    """
+    Serializer for DukeDS users API
+    """
+    id = serializers.UUIDField()
+    username = serializers.CharField()
+    full_name = serializers.CharField()
+    first_name = serializers.CharField()
+    last_name = serializers.CharField()
+    email = serializers.CharField()
+
+    class Meta:
+        resource_name = 'duke-ds-users'
+
+
+class DDSProjectSerializer(serializers.Serializer):
+    """
+    Serializer for DukeDS projects API
+    """
+    id = serializers.UUIDField()
+    name = serializers.CharField()
+    description = serializers.CharField()
+
+    class Meta:
+        resource_name = 'duke-ds-projects'
